example, resolve: synchronous routes and promises
Added support for synchronous routes and promises. If the resolver
function does not explicitly list the callback as an argument then
it is assumed to be synchronous. Guards are put in place to make
sure that the callback is only called once and that errors are
caught. Updated the example to use asynchronous routes only when
necessary.