example, location: added location to express
Added the ability to resolve a location from express, which will be
assigned to the `Location` header of the response. This is currently the
only response header that is capable of being assigned. Eventually, it
would be nice to provide a function for supplying a header name and type
to resolve. Also updated the example to showcase the new functionality.