given: add givens to cache
Updated ziploc to keep a list of the given types and values separate
from the resolvers to ensure that they have a higher priority. The
givens are added to the resolve cache before resolving to make sure that
the resolver functions are not used in their place. This also has the
added benefit of no longer requiring the routes to be scored and sorted,
thus making the resolve function much faster.